move buffer functions into GopherCAN.c
