+++ GopherCAN.c
@@ -4,7 +4,6 @@
 //  return request from other modules and call CAN commands
 
 #include "GopherCAN.h"
-#include "GopherCAN_ring_buffer.h"
 #include "GopherCAN_network.h"
 
 // static function prototypes
@@ -19,6 +18,10 @@ static S8   service_can_rx_message_std(CAN_MSG* message);
 static S8   service_can_rx_message_ext(CAN_MSG* message);
 static S8 encode_parameter(CAN_INFO_STRUCT* param, U8* data, U8 start, U8 length);
 static S8 decode_parameter(CAN_INFO_STRUCT* param, U8* data, U8 start, U8 length);
+static void init_buffer(CAN_MSG_RING_BUFFER* buffer, CAN_MSG buffer_memory_ptr[], U8 buffer_size);
+static void remove_from_front(CAN_MSG_RING_BUFFER* buffer);
+static void add_message_by_highest_prio(CAN_MSG_RING_BUFFER* buffer, CAN_MSG* message);
+static void copy_message(CAN_MSG* source, CAN_MSG* dest);
 
 #if NUM_OF_BUSSES > 1
 static CAN_MSG_RING_BUFFER* choose_tx_buffer_from_hcan(CAN_HandleTypeDef* hcan);
@@ -1045,6 +1048,95 @@ static S8 send_error_message(CAN_ID* rx_id, U8 error_id)
 
 	// send the CAN message
 	return tx_can_message(&message);
+}
+
+
+// init_buffer
+//  initialize the buffer with the values passed in as parameters
+void init_buffer(CAN_MSG_RING_BUFFER* buffer, CAN_MSG buffer_memory_ptr[], U8 buffer_size)
+{
+    buffer->head = 0;
+    buffer->fill_level = 0;
+    buffer->size = buffer_size;
+    buffer->message_buffer = buffer_memory_ptr;
+}
+
+
+// remove_from_front
+//  will remove the first element of the ring buffer. If the buffer is empty it will do nothing
+static void remove_from_front(CAN_MSG_RING_BUFFER* buffer)
+{
+    // don't do anything if the buffer is empty
+    if (IS_EMPTY(buffer))
+    {
+        return;
+    }
+
+    // move the head to the next element
+    buffer->head = (buffer->head + 1) % buffer->size;
+
+    // decrement the fill level
+    buffer->fill_level--;
+}
+
+
+// add_message_by_highest_prio
+//  This function will add message to the buffer based on the ID of the message. Higher
+//  priority messages (lower ID) will be towards the front, with lower priority
+//  messages (greater ID) will be towards the back. Removing from the front will get
+//  the highest priority message. This function assumes the buffer is not full
+void add_message_by_highest_prio(CAN_MSG_RING_BUFFER* buffer, CAN_MSG* message)
+{
+    CAN_MSG* buffer_message = GET_FROM_BUFFER(buffer, 0);
+    S16 c;
+
+    // start from the back of the buffer, moving each message towards the back
+    // by one and put the new message in the correct spot by ID. If the buffer
+    // was empty when the message first went through here, it will put the new
+    // message in position 0
+    buffer->fill_level++;
+    for (c = buffer->fill_level - 2; c >= 0; c--)
+    {
+        buffer_message = GET_FROM_BUFFER(buffer, c);
+
+        if (
+            (message->header.IDE == CAN_ID_STD &&
+            buffer_message->header.IDE == CAN_ID_EXT)
+            ||
+            (message->header.IDE == CAN_ID_EXT &&
+            buffer_message->header.IDE == CAN_ID_EXT &&
+            message->header.ExtId >= buffer_message->header.ExtId)
+            ||
+            (message->header.IDE == CAN_ID_STD &&
+            buffer_message->header.IDE == CAN_ID_STD &&
+            message->header.StdId >= buffer_message->header.StdId)
+        ) {
+            // new message is lower priority, insert behind this buffer message
+            buffer_message = GET_FROM_BUFFER(buffer, c + 1);
+            break;
+        }
+
+        // move this message back by 1 and try again
+        copy_message(buffer_message, GET_FROM_BUFFER(buffer, c + 1));
+    }
+
+    // put the message into the buffer at this position
+    copy_message(message, buffer_message);
+}
+
+
+// copy_message
+//  function to copy all of the data in source to dest by value, not by refernce
+static void copy_message(CAN_MSG* source, CAN_MSG* dest)
+{
+    U8 c;
+
+    dest->header = source->header;
+
+    for (c = 0; c < dest->header.DLC; c++)
+    {
+        dest->data[c] = source->data[c];
+    }
 }
